spark/requirements.txt
pyspark==3.5.3
pypdf==4.3.1
qdrant-client==1.9.1
sentence-transformers==3.3.1
requests==2.32.3
python-dotenv==1.0.1
